fix!: Move .expand to Compiler
to reduce cyclic dependency
