+++ src/library/extra_lib.js
@@ -288,7 +288,7 @@ var macroexpand_1 = function(x){
   return x;
 }
 define_syntax("%macroexpand", function(x){
-  var expanded = Interpreter.expand(x.cdr.car);
+  var expanded = Compiler.expand(x.cdr.car);
   return List(Sym("quote"), expanded);
 });
 define_syntax("%macroexpand-1", function(x){
@@ -297,7 +297,7 @@ define_syntax("%macroexpand-1", function(x){
 });
 
 define_libfunc("macroexpand", 1, 1, function(ar){
-  return Interpreter.expand(ar[0]);
+  return Compiler.expand(ar[0]);
 });
 define_libfunc("macroexpand-1", 1, 1, function(ar){
   return macroexpand_1(ar[0]);
